requirements.txt
requests
user_agent
pystyle
colorama
pyyaml
